WIP: Change over forgeZarith to not round trip through hex

+++ src/Tezos/Forging.h
@@ -12,4 +12,4 @@ using namespace TW;
 Data forgeBool(bool input);
 Data forgePublicKeyHash(const std::string &publicKeyHash);
 Data forgePublicKey(PublicKey publicKey);
-Data forgeZarith(int input);
+Data forgeZarith(uint16_t input);
